+++ src/main/java/com/wanted/preonboarding/ticket/presentation/controller/QueryController.java
@@ -27,7 +27,6 @@ public class QueryController {
      */
     @GetMapping("/all/performance")
     public ResponseEntity<ResponseHandler<List<ResponsePerformanceInfo>>> readAllPerformances() {
-        log.info("QueryController getAllPerformanceInfoList");
         return ResponseEntity
             .ok()
             .body(ResponseHandler.<List<ResponsePerformanceInfo>>builder()
@@ -36,20 +35,6 @@ public class QueryController {
                 .data(ticketSeller.readAllPerformances())
                 .build()
             );
-    }
-
-
-    /**
-     * 예약 조회 시스템
-     * Request Message: 고객의 이름, 휴대 전화
-     * Response Message: 예매가 완료된 공연의 정보(회차, 공연명, 좌석정보, 공연ID) + 예매자 정보(이름, 연락처)
-     * @param dto
-     * @return
-     */
-    @GetMapping("/customer/reservation")
-    public ResponseEntity<BaseResDto> readReservation(@RequestBody RequestReserveQueryDto dto) {
-        log.info("QueryController getReserveInfo");
-        return ResponseEntity.ok(ticketSeller.getReserveInfoDetail(dto));
     }
 
 }
