+++ src/main/java/com/wanted/preonboarding/ticket/presentation/controller/ReservationController.java
@@ -2,6 +2,7 @@ package com.wanted.preonboarding.ticket.presentation.controller;
 
 import com.wanted.preonboarding.ticket.application.TicketSeller;
 import com.wanted.preonboarding.ticket.aop.dto.BaseResDto;
+import com.wanted.preonboarding.ticket.domain.dto.request.ReadReservationRequest;
 import com.wanted.preonboarding.ticket.domain.dto.request.CreateReservationRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,6 +27,19 @@ public class ReservationController {
     @PostMapping
     public ResponseEntity<BaseResDto> createReservation(@RequestBody CreateReservationRequest createReservationRequest) {
         return ResponseEntity.ok(ticketSeller.createReservation(createReservationRequest));
+    }
+
+
+    /**
+     * 예약 조회 시스템
+     * Request Message: 고객의 이름, 휴대 전화
+     * Response Message: 예매가 완료된 공연의 정보(회차, 공연명, 좌석정보, 공연ID) + 예매자 정보(이름, 연락처)
+     * @param readReservationRequest
+     * @return
+     */
+    @GetMapping("/customer")
+    public ResponseEntity<BaseResDto> readReservation(@RequestBody ReadReservationRequest readReservationRequest) {
+        return ResponseEntity.ok(ticketSeller.readReservation(readReservationRequest));
     }
 }
